Fix typo in cell
This commit addresses a typo in the code that yields unexpected results.

In the cell, `results, run_simulation(...)` causes the simulation to be
run correctly, but the results are never assigned back to the original
`results` variable, meaning that the fraction of infected individuals is
never updated over the course of the sweep. When running this code, it
then seems like the fraction of infected individuals never changes,
regardless of the number of doses (and resulting changes in beta).